trafficserver: switch querybuilder scholarly and main to wikikube
This change switches query service /querybuilder endpoint to
wikikube Kubernetes as well. The query-gui is running smoothly
so let's switch querybuilder as well.
The dedicated mapping for /querybuilder is removed so it falls
back to the default, which is miscweb.discovery.wmnet:30443.

First test on query-main and query-scholarly. After that we can continue
with query.wikidata.org.

Bug: T350793
Change-Id: If4d084be74c75800be02ad75d3fe40bb55642f7d