hiera: acme_chief: add wikimedia-ech.org
Per the ECH RFC [0], a client expects to see a valid certificate for
public_name, or the outer SNI. This commit fetches certificates for
wikimedia-ech.org, our standard public_name that will encrypt all inner
SNIs. ("The client MUST verify that the certificate is valid for
ECHConfig.contents.public_name.")

We are only doing this for the durum* hosts, so restrict to them.

[0] - https://datatracker.ietf.org/doc/draft-ietf-tls-esni/

Bug: T205378
Hosts: acmechief1002.eqiad.wmnet
Change-Id: I6b5e7ca12f0095ef1f6cef744b2cc2375feb754b