Add a canonical list of sensitive LDAP groups
Right now we need to track sensitive files in multiple locations which
is error-prone (wikitech, offboard script, cross-check script).

Instead let's create one canonical list and then rework the tooling
to use it. The format is a standard config/ini file, so that we
can flexibly add new attributes. The ones currently added are:
- sync_to_posix: Specifies the POSIX group to sync a user against
                 (not yet implemented, but requested in T375729
- bitu: Indicates whether that group is requested in Bitu (if the
        attribute is absent, it's not). This also serves as a reminder
	of which groups still need to be migrated

Change-Id: I979785753e06bc40f7c0663bd5ede6ec180257d7