Create insetup role for Data Persistence with nftables and rename existing one
Currently all our insetup roles default to Ferm, but with an increasing
amount of roles defaulting to nftables this is becoming problematic:
It means these are installed with Ferm initiaally and then Puppet
applies the switch to nftables. But this switch requires a reboot (to
unload the iptables kernel modules), which might be missed, causing
issues later on.

As such to make the choice of the firewall stack part of the role, so
that it can be explicitly selected when creating the racking task
for DC ops:
- Reanem the existing insetup roles for Data Persisntence to
  insetup::data_persistence_ferm
- create a new role insetup::data_persistence_nftables which defaults to
  nftables (currently no DP role defaults to nftables)
- rename existing entries in site.pp depending on role config

Bug: T389825
Change-Id: I13b50df5c554dbaed2d96857972d489e43d5f737