Create insetup role for Traffic with nftables and rename existing one
Currently all our insetup roles default to Ferm, but with an increasing
amount of roles defaulting to nftables this is becoming problematic:
It means these are installed with Ferm initially and then Puppet
applies the switch to nftables. But this switch requires a reboot (to
unload the iptables kernel modules), which might be missed, causing
issues later on.

As such to make the choice of the firewall stack part of the role, so
that it can be explicitly selected when creating the racking task
for DC ops:
- Rename the existing insetup roles Traffic to insetup::traffic_ferm
- create a new role insetup::traffic_nftables which defaults to
  nftables
- The only existing insetup::traffic entry is actually wrong, as
  a cp host it should use the noferm variant, fixing that along.

Bug: T389825
Change-Id: I41c3d36ed133af10e8eb6c9f00d3a224c2a6a53e