Create insetup role for WMCS with nftables and rename existing one
Currently all our insetup roles default to Ferm, but with an increasing
amount of roles defaulting to nftables this is becoming problematic:
It means these are installed with Ferm initially and then Puppet
applies the switch to nftables. But this switch requires a reboot (to
unload the iptables kernel modules), which might be missed, causing
issues later on.

As such to make the choice of the firewall stack part of the role, so
that it can be explicitly selected when creating the racking task
for DC ops:
- Rename the existing insetup role for WMCS to insetup::wmcs_ferm
- create a new role insetup::wmcs_nftables which defaults to
  nftables
- Move the existing insetup hosts to the variants using ferm or
  nftables based on the setup for the current roles in prod
- The entries for cloudvirt* and cloudnet* were actually wrong, these
  should use the noferm variant since they don't use Ferm. This patch
  also corrects this.

Bug: T389825
Change-Id: Icfe441ba212a39a13f69b63a83b5ef43f9091a34