search: Transparently retry 503 errors
We see a low level of connection errors to the search
infrastructure. On a percentage basis they are very low, but it
would be nice to avoid them. These are most commonly of the form:

  Status code 503; upstream connect error or disconnect/reset
  before headers. reset reason: connection failure

The retry is limited to 503's, as 500's are expected for certain
query errors such as invalid regex's provided by users.

Bug: T390612
Hosts: deploy1003.eqiad.wmnet
Change-Id: Idc0574709e8814724860cffe7b3760745015cb4b