Revert^4 "cache::backend: Switch mw-wikifunctions to ingress"
This reverts commit 2ab89798916c4c74423ca9401c443dfdd58fd0e7.

Reason for revert: We figured out that the cause was ATS sharing connections for
the same IP and behind the wikikube-ingress IP are multiple services. In a non
deterministic way, some requests would end up at the wrong backend returning 404s.

This patch also removes the sharing of connections for the same destination IP
for wikifunctions.

Bug: T384944
Bug: T390854
Change-Id: Ifd811d3d51d61d7090e439884daddc8d9e8635d3