install-server: also run configure_swift_disks for apus-*
Apus nodes, like the moss nodes that came before, are set up to use
partman/custom/cephadm_raid1_leavelvm.cfg
for preseeding.

That relies on partman_early_command.sh to select the two disks to
install on (in the function configure_swift_disks ), so we need to
make sure that is run on apus-* nodes as well as moss-* nodes.

Bug: T390578
Change-Id: I5930a85b0ad589fb0323d21ecaf8032126d4715e