mw::periodic_jobs: Pass command through untouched
Why:
Passing the command as a whitespace split array to the final yaml poses
"interesting" challenges as it goes through three layers of templating.

What:
Just pass down the command as a quoted string to be appended to the args
array of the final kubernetes CronJob.

Hosts: mwmaint1002.eqiad.wmnet, deploy1003.eqiad.wmnet
Bug: T341555
Change-Id: Id6f4067312dfde0bd9c7c1acdf5862984c92f7bc