Create insetup role for ServiceOps with nftables and rename existing one
Currently all our insetup roles default to Ferm, but with an increasing
amount of roles defaulting to nftables this is becoming problematic:
It means these are installed with Ferm initially and then Puppet
applies the switch to nftables. But this switch requires a reboot (to
unload the iptables kernel modules), which might be missed, causing
issues later on.

As such to make the choice of the firewall stack part of the role, so
that it can be explicitly selected when creating the racking task
for DC ops:
- Rename the existing insetup role for ServiceOps to insetup::serviceops_ferm
- Create a new role insetup::servieops_nftables which defaults to
  nftables
- Move the existing insetup hosts to the variants using ferm or
  nftables based on the setup for the current roles in prod

Bug: T389825
Change-Id: I8a4411074ec7190933573e11cdbeae945963d06a