hiera: cleanup gitlab-runner docker gc settings
To start somewhere with the Hiera cleanup described in T390948
pick the docker garbage collection settings.

The interval and timeout values at the host level are exactly the
same as on puppet role level and cloud-wide level.

If there is no difference then it makes no sense to have specific
settings at the host level, so remove them.

While there is a case to be made that theoretically other projects
could use these values from cloud.yaml it seems very unlikely there
would be other puppet roles than the existing ones for gitlab runner.

So the puppet role level seems to be the best fit.

Bug: T390948
Change-Id: Ia644bdc088675cd04a349467645c4674a3dadaef