hiera: cleanup some gerrit and etherpad hiera values
More Hiera cleanup for collab cloud services.

The etherpad service ensure is already set to running
on the puppet role level and specifically to stopped
on a host name level.

There seems to be no reason to also set a default for all
of cloud and clutter up cloud.yaml.

The only situation where this would be helpful was if someone
started writing new puppet roles for etherpad that aren't the
production role.

The gerrit IPv6 setting is already on the devtools project
level and I am not aware of any gerrit instances outside of devtools.

The gerrit monitoring is set to true on the puppet profile class
and to false on the devtools project level. Again, repeating it
another time for all of cloud would only have any benefit if there
were other projects with gerrit in them.

But even the one gerrit test instance we have in devtools is shut down.

Bug: T390948
Change-Id: I315e1051694578109c70acab4c63766180c81277