cirrussearch: add puppet 7 hieradata to DC-specific config
The reimage cookbook doesn't seem to pick up the config
that enables Puppet 7 for the cirrus::opensearch role,
which results in the cookbook stalling at
'Nagios_host resource with title cirrussearch2056 not found yet'

We worked around this with our original host (cirrussearch2055)
by adding directly to host-specific hieradata, but we'd prefer
to avoid doing that for each host. Try with DC-specific config
to see if it makes a difference.

Hosts: cirrussearch2055.codfw.wmnet
Bug: T388610
Bug: T383811
Change-Id: I465142feff6b6dafd0c99aff55e027b68578fca6