Add sni_rewrites_host_header
Why:
Up to now, we always assumed that if we were setting SNI, we also wanted
to use the SNI value as the HTTP Host header. This made sense for all
the interal services, however in a world of a service being able to
answer to more than 1 virtual hosts, e.g. MediaWiki, this no longer
makes sense.

What:
Add an sni_rewrites_host_header toggle for service mesh

Hosts: auto
Change-Id: Iaf8f0173ab3dfe966e9db4d0f50fcccad0de4607