mjolnir: temp remove msearch daemon from codfw
The codfw search cluster is currently in a mixed state with
elasticsearch and opensearch instances. This introduces a
variation in feature collection that we aren't yet prepared to
handle.

For now remove the search loader from codfw and let eqiad, which
is only running elasticsearch, handle feature collection as
normal. Once codfw is all opensearch we will want to flip things
the other way and collect only from codfw.

Hosts: search-loader1002.eqiad.wmnet,search-loader2002.codfw.wmnet
Change-Id: I9ef46c733480b1da6b50df76f305164edaa620d3