cumin: Update insetup role report
Why:
With the introduction of the new specialized insetup roles for ferm and
nftables, the insetup report would not work anymore.

What:
Update the report to use the new role names.
Fix also a couple of roles that wouldn't match.
Re-enable the report on cumin1002, as apparently its setting was not
migrated from cumin1001.

Hosts: auto
Bug: T389825
Change-Id: Ia67e8d29a76fb77ef0ffd715f51b91f34b7910f1