trafficserver: switch querybuilder scholarly to wikikube
This change switches query service /querybuilder endpoint to
wikikube Kubernetes as well. The query-gui is running smoothly
so let's switch querybuilder as well.
The dedicated mapping for /querybuilder is removed so it falls
back to the default, which is miscweb.discovery.wmnet:30443.

Migration start with query-scholarly. After that we can continue
with query and query-main. Second try of
If4d084be74c75800be02ad75d3fe40bb55642f7d.

Bug: T350793
Change-Id: Ibf2d067b70cb0b352759a3fc194f4980dc9cfd1c