hieradata: add phab shell user groups on phab1005
The host phab1005 does not have a role yet but we are going to use
it to test a phabricator upgrade involving a db schema change.

So the regular users with phab host shell access need to have access
to it and we can start with that before applying the full phabricator
role. This way we can share the DB password in user home dirs.

Bug: T390034
Bug: T377889
Change-Id: Ia09d1767ed793856a12c2ff3c58632e638f5d616