phabricator::migration: use nftables as firewall provider
For the special phabricator::migration role/profile, which
we only use every once in a while, but comes in handy again,
ensure it uses nftables and not ferm as firewall provider.

This is matching the phabricator production role
and is to minimize the diff when a new machine transitions
from the new insetup::collaboration_services_nftables role.

Bug: T370677
Bug: T377889
Change-Id: Ib7a31fe524dffa98d9458e0d8e8ebea52357bfcb