zookeeper: onboard an-conf1004 to the cluster
As part of the an-conf hosts refresh we shall onboard one host at a time to the zk cluster, expand
the cluster to 6 hosts then reduce it back to 3 hosts. This change onboards an-conf1004 to the
zk cluster. the service will need to be restarted  in order to pickup the new host.

Bug: T374922
Change-Id: I058b554b176735700804e843dd8843b912a5ab94