zookeeper: onboard an-conf1005 to the cluster
As part of the an-conf hosts refresh we shall onboard one host at a time to the zk cluster, expand
the cluster to 6 hosts then reduce it back to 3 hosts. This change onboards an-conf1005 to the
zk cluster. the service will need to be restarted  in order to pickup the new host.

Bug: T374922
Change-Id: Iad3fb55d2fbccfd816db34245138660d5aae8a48