zookeeper: onboard an-conf1006 to the cluster
As part of the an-conf hosts refresh we shall onboard one host at a time to the zk cluster, expand
the cluster to 6 hosts then reduce it back to 3 hosts. This change onboards an-conf1006 to the
zk cluster. the service will need to be restarted  in order to pickup the new host.
This is the last of the new hosts, next shall be removing the hosts from the cluster, this shall
be done once all services are running on the new hosts.

Bug: T374922
Change-Id: I0eda533539d080004bf18afaa4870c6dcc9b0f9e