Add an rsync fragment to permit dse-k8s pods to sync mediawiki dumps
We are migrating the mediawiki-dumps-legacy to Airflow and Kubernetes,
but we will need an updated rsync configuration to permit the dse-k8s
worker pods to sync their data files and metadata files to the
clouddumps servers.

This commit adds a new rsync config fragment that will use password
authentication instead of a simple hosts_allow list.

The password will then be added to the sync-utils pods that run as part
of the Airflow DAG for generating the dumps.

Hosts: clouddumps1001.wikimedia.org,clouddumps1002.wikimedia.org
Bug: T389784
Bug: T390738
Change-Id: I97e92c383307728c0b5ad746bd5d39d8ee523a2f