Profile::Mediawiki_deployment: add 'dir' field
The `dir` field was added in [0] and will be used to configure the
sub-path within helmfile.d where services are expected to be defined.

Example: In order to find the mediawiki-dumps-legacy namespace-level
service directory in helmfile.d/dse-k8s-services, one would set:

  - namespace: mediawiki-dumps-legacy
    dir: dse-k8s-services
    [ ... snip ... ]

In the absence of a `dir` field, helmfile.d/services (scap default) is
used.

[0] https://gitlab.wikimedia.org/repos/releng/scap/-/merge_requests/681

Bug: T388761
Change-Id: If3d0499670094641bce6957f6aab1e1ca3de613b