logstash: transform err field usage to hash
Why: Regularly, one application declares the err field to be a
string when most usage expects err to be a hash.  This is a race
condition affecting the indexing ability for the entire day's index.

What: Mutate all usage of the err field to a hash.  If it is a string,
move the contents into err.message.

Hosts: logstash1023.eqiad.wmnet
Change-Id: Ibb9bf938e3b3f79c88944d39508aae6f9672754b