phabricator::migration: add scap::target, add deploy scripts, rm symlink
Attempting to fix:

[deploy1003:~] $ scap install-world -l phab1005.eqiad.wmnet
Aborting: No hosts to install. Nothing to do

Which seems to mean we need scap::target on this machine.

But adding that creates a duplicate definition with the existing
symlink we used to create to the scap binary.

Additionally, add the deployment scripts that appear in sudo rules
and should be needed for an initial deploy.

Somehow we did not need this stuff last time, but I guess
something changed since we migrated a phab host last time.

Bug: T377889
Change-Id: I33f97c87b07b4e83dc894105f343873e4f474386