trafficserver: generalize mw-next-routing.lua and prep for PHP 8.3
The primary goal of this change is to make mw-next-routing.lua easier to
use in future migrations, including yearly PHP migrations, or really any
other migration that might benefit from using the "next" deployments in
a similar way (i.e., driven by the presence of a specific pattern in the
Cookie header).

Speficially:
1. References to the 8.1 migration are removed, and the specific Cookie
   pattern for identifying eligible requests is now a config field.
2. The tests now better reflect this separation, and have also been
   restructured to use partially table-driven tests for exhaustive
   checks (e.g., verifying behavior for all discovery services).
3. The config is extended with an `enabled` flag, which can be used to
   early-exit the script when not in use, opening the door to leaving it
   in the relevant plugin chains at steady-state.

I briefly considered introducing a more powerful mechanism to directly
define the enrollment logic in config, but I could not figure out a way
to render it trivially safe. Plus, pushing logic into configuration has
the added downside of discouraging writing proper tests.

Ultimately, constraining configuration to specifying only the eligible
Cookie pattern seems like the right tradeoff on that front.

Finally, in advance of the PHP 8.3 migration, the production config is
updated to target PHP_ENGINE=8.3, although with the enabled flag set to
false (and load_fraction set to 0) for now.

Bug: T391421
Change-Id: I5c84a174b12ba38366956860acc621c86435901e