Revert "mjolnir: temp remove msearch daemon from codfw"
This reverts commit 1a34a58fae5af22614a68b4b834d30b9a17d8758.

Reason for revert: We identified the problem as being related to
missing values handling in the 1.x version of the LTR plugin. We
released a new version of the plugin, 1.5.3-wmf1-os1.3.20, which
backports the fix found in opensearch 2 and makes 1.x behave the
same as elastic 7 and opensearch 2. With the updated plugin we no
longer need to worry about the mixed cluster.

Change-Id: I3542e37f57ac82d5feb26a75e66d4aa983289c54