openstack: Tidy up wmcs-wikireplica-dns script
The zones managed by this script are now all in the "cloudinfra"
openstack project, so we can simplify the logic.

I removed the check for "supported" zones, we can trust the yaml file to
contain zones that we want to update, and in any case this will only
update zones in the "cloudinfra" project.

I also took the chance to modify the logging, increasing the default
verbosity to print a little more info when the script is running.

Bug: T374953
Change-Id: If7111e9e8dddaba83b3349280b76f50728eac722