miscweb: remove query-service from legacy vms
All query-gui and querybuilder components were migrated to wikikube.
This change removes the query components from the legacy miscweb VMs.

Before merging this change we could do a less disruptive a2dissite
which is easier to rollback in case something uses the old endpoint.

This change also adds monitoring for the Kubernetes service.

Bug: T350793
Change-Id: Ie73a5cbff7d8d913db0ba71b05b358d9cf378082