idp-test: disable monitoring notifications, copy theme setting
The host "idp-test2005" has been Icinga alerting on IRC for basically a
full week that it was out of disk.

And it was 100% full.

I ran an "apt-get clean" which brought it back down to 94% usage.

Two points here:

a) Hosts that have "test" in the name probably should not have any
monitoring or at least no alerting. So copying from other test hosts
and disabling the notifications.

b) If this was not a test host it still seems nobody would have been
effectively notified of this because "IRC only" alerts are not working
anymore. So it should likely be converted to notification by email or
task.

Additionally I noticed the existing host Hiera file for idp-test in
eqiad and copied it over to make eqiad and codfw consistent.

Assuming that theme setting makes sense for both.

Change-Id: Ia3fb6abed37d9b7f03f9ada9ad90963d147069c3