dynamicproxy: Delete dns entries before removing proxy records
I have an edge case where somethign is crashing in the middle of removing
a record. When that happens after the proxy record has been removed but before
some or all dns records are removed, the dns records are orphaned.

With this the other way around, the proxy record is still present
after a failure which allows us (or tofu) to re-attempt the deletion.

Bug: T391718
Change-Id: Ife5dfadaf1f54df7dd9d2429234f30594bdbf196