miscweb: remove static-rt profile from legacy miscweb role
We are not going to create a static-rt site after all.

Instead we are keeping scraped HTML files and a database
dump and are storing it in a non-public dumps location.

It would be possible to store those HTML files on some
webserver anytime in the future though.

But for now we want to decom the legacy miscweb machines.

Bug: T385777
Bug: T300171
Change-Id: Ide295c4a059ffa7ff22e770c8c8c468873b76c14