hieradata: switch mw-script main release to PHP 8.1
The switches the mw-script main release to PHP 8.1 and stops maintaining
the release values file for the next release.

The following patch in this series will remove 7.4 fallback support in
mwscript-k8s, leaving only 8.1 support referencing main.

Note that this does not result in *deletion* of the next values file, so
it's safe to apply this (via scap) before merging the following patch.

Bug: T391057
Change-Id: I5d7fd6156a07a0d48997968fa60a7b06d099d970