phabricator/scap: disable scap bootstrapping on phab1005
Host phab1005 is not in production yet and in the process of
being setup.

There is a scap bootstrapping issue because the scap home dir
is hardcoded.

See I64b876552a27b0f4f for details and a suggested fix.

But because that touches all of scap I am first disabling
bootstrapping on this one host to not leave a broken puppet run
until we get to fix it.

Bug: T377889
Bug: T392386
Change-Id: Ie27ba1d4be2569d3ee8a14ffde77867b7776e1ae