gerrit: add nftables rule to allow Istanbul Hackathon hotel network
This is the IPv4 network given to us as the one of the hotel in Istanbul
where the Hackathon is going to happen in early May.

Still tbd if there is an IPv6 as well and going to do this for GitLab
separately as well.

Tried to find the easiest fix for a temporary situation so did not
try to introduce a whole new set to be allowed.

It's supposed to be reverted after the event.

Order does not seem to matter in this case but I set it to 11 to come
after the abusers drops.

Bug: T382309
Change-Id: I7f1e6cf69ece12d3b9eb5fb50922fc3b68fd66fc