firewall/nftables_throttling: temp add rule to allow Istanbul Hackathon
This is the IPv4 network given to us as the one of the hotel in Istanbul
where the Hackathon is going to happen in early May.

Additionally there was a single IP for the wifi network that was given
to me.

Tried to find the easiest fix for a temporary situation so did not
try to introduce a whole new set to be allowed.

It's supposed to be reverted after the event.

Also see:  I005c9e921b0752 (MediaWiki), I7f1e6cf69ece12 (Gerrit)

Bug: T382309
Change-Id: Iee5d6be006da12e097542908e41702975d77ce3f