Add krb1002 to kerberos_kdc_servers
The packages are properly installed, so as a next step we can add it to
the server list to setup replication of data to new new node.

krb1002 is not yet added to kerberos_kdc_servers_to_clients, so it won't
be queried yet.

Bug: T390863
Hosts: krb1001.eqiad.wmnet,krb1002.eqiad.wmnet,krb2002.codfw.wmnet
Change-Id: Idec5ee734b9586f74f494845b8fc8d8b3f3a9a17