hieradata: Expand GitLab blocklist for new WMCS IP space
See https://wikitech.wikimedia.org/wiki/Help:Cloud_VPS_IP_space. This
expands the existing VLAN /21 to the /17 that covers it and the VXLAN
subnets replacing it.

The v4 public space has not changed. There's a new v6 range, but AFAIK
the GitLab runner Docker setup is v4 only and the setting only allows a
single subnet, so I'll leave adding that for later.

As a side note, it'd be great to source this from Puppet instead of
duplicating it here.

Change-Id: Id424d44c83942c248de8aaa1ab9786ea822fab85