P:auth: temporarily skip returning a WARN on check_authdns_state
This causes the restart cookbook to fail since the host is not pooled.
Temporarily set a depooled host to return OK; we will revert this when
done.

The correct fix for this is in the cookbook itself where it can skip
such services but since Spicerack does not support that, we do this
temporary fix for now.

Change-Id: I00ac79f0de1b8c23160764ef8e4c5faa0bd3c87f