P:durum: only log ECH status for ECH-enabled clients
There is no reason to turn on logging for all other bits of durum as we
had disabled that on purpose. To simplify this, simply log the ECH
status for ECH-enabled clients and disable logging for everywhere else.

This also helps avoid check_python spamming the access logs.

Hosts: durum3003.esams.wmnet, durum1001.eqiad.wmnet
Bug: T205378
Change-Id: I26b25d3e3988639c91ce9206005896d502f3517e