query-legacy-full: set cluster in hiera
This should allow us to view query-legacy-full's metrics
separately from the other full graph hosts.

We apparently also need a corresponding entry in the
wikimedia_clusters hieradata value, so add that.

This service is not behind LVS so we're hoping we
don't need it to be in LVS for a wikimedia_clusters
entry to be added.

Hosts: wdqs2009.codfw.wmnet
Hosts: O:prometheus
Bug: T384422
Change-Id: Id089acf1d41de28b75c58f30c6c5bb9d2d0350dd