gerrit: split Gerrit and Gitiles proxy pools
Apache proxy module comes with two built-in workers: one for forwarding,
the other for reverse proxy). They are configured with the ProxyPass
directive.

Reorder the ProxyPass directive to have two proxy worker pools, this
will let us set different settings on each working pools.

Add a more specific ProxyPass for /r/plugins/gitiles/ which is proxyed
as /g/. Those two are routed to the same backend URL and thus forming a
worker pool dedicated to Gitiles. I have tested locally this work, but
that would need to be tested once deployed.

Later we can then set different setting for the Gitiles pool.

Bug: T392467
Change-Id: I6ce0043d084c8362511fcd637a24ef6ffe6b4a29