Update clouddumps contactgroups to reflect shared ownership
The clouddumps servers are important to the members of the Data Platform
teams as well as to users of cloud services. For instance, the NFS
exports from these servers are mounted on stat hosts and an-launcher
hosts in order to import the generated dump files to HDFS.

Since the Data Platform team also has responsibility for the generation
and publishing of the dumps files, we should update the contact group
information for the hosts to reflect this.

Hosts: O:dumps::distribution::server
Change-Id: I901c735c3509e141a033bea2256a988715a123de