Preseed: select manual setup for apus-be[1,2]004
These new backends will have their SSDs in a hardware RAID-1 on a boss
card; we don't know how this will appear to the operating system. It
_might_ just show up reliably as /dev/sda, in which case we could just
use flat.cfg, but more likely we'll have to do something a little more
fiddly to determine which drive to use - for future reimages, we want
to be confident of not damaging data on the NVME or spinning disks,
which are used by Ceph.

Bug: T392844
Bug: T392845
Change-Id: Ia86488358a53ead68197962087106ef717c6b8d6