Enable the dumpsgen user to use an rsync server over ssh from dse-k8s-eqiad
We need to allow pods running on the dse-k8s cluster to eqiad to send
the dumps output files to the clouddumps servers for distribution.

This commit adds an authorized_keys file to permit this, as well as
opening up the SSH port to the dse-k8s pod IP ranges.

We force the command that is used to be rsync in server mode, without
the --delete option, since the clouddumps servers already have
housekeeping configuration to clean up old dumps.

Hosts: O:dumps::distribution::server
Bug: T390738
Change-Id: I3d5388fa025c91d0c1f57c97d5d4a3ac31e0cf30