P:durum and hiera: update health check path
For the anycast-hc health checks, use the dedicated /check path instead
of the blanket /. This is more specific and also allows us to prepare
for turning on selective logging for ECH later.

This will allow us to continue logging ECH stuff but just that and
nothing else.

Hosts: durum1001.eqiad.wmnet, durum3003.esams.wmnet
Change-Id: Icba4fbaeac8946fb9dfc39d780b5a739b7764f96