P:mediawiki::php: add uuid extension for PHP 8.1+
Although the uuid extension is not yet used, we anticipate that it will
be soon.

As of now, while we *could* support the extension on 7.4, that is not
currently planned given the limited longevity thereof. Thus, it's only
installed on 8.1 (or later).

Note that this is already present in the php8.1 production images.

Hosts: P:mediawiki::php
Bug: T373752
Change-Id: Ibb39e2252c56dfd35c4a32555b44d0bd499dcf84