Stop passing krb2002 to Kerberos clients
Now that we have two active KDCs (krb1001, krb1002), drop krb2002 from
the list of KDCs used by Kerberos clients. This will allow us to upgrade
krb2002 to Bookworm next. The data layout of Kerberos isn't very friendly
to reimages and the server is up for renewal in two years anyway, so this
is one of the few cornercases where we'll rather dist-upgrade.

Bug: T390863
Change-Id: Ib8ba542f47da65b554993726c7366f329e52681a