mw::maintenance: migrate all remaining general updatequerypages jobs
Migrate two large jobs and one single-shard jobs. These jobs are quite
high priority, but there is little reason that these jobs will be
impacted by the migration when all of the scripts run in the same
manner. They run twice a month, and so we are unlikely to hold until
after their next run to migrate. The other forms of these jobs are
running successfully.

Bug: T388534
Change-Id: I322db28cff5a880dbf763281dbd4178b66919177