gerrit: enable backups on gerrit2003
During an incident in a gerrit-failover we noticed
we don't have enough disk space to both keep a local
backup of gerrit data and also restore from Bacula.

We wanted to save a backup of both gerrit1003 and gerrit2002
somewhere and then copied them over to new machine gerrit2003.

This would start Bacula backups also on this new machine, which
would pick up the files from the other 2 machines which could
later be deleted again.

Bug: T393034
Change-Id: I71222f69cf56c9bc9fef21d7a62794544452eedb