gerrit: add a second replica, start replicating to gerrit2003
So far we always just had a single primary server and single replica
host for gerrit.

Then we added a third machine, new hardware gerrit2003 which is not
in production yet.

This is to start replication to this new machine in addition with
the goal of permanently having 2 replicas.

I suggest to follow the gitlab naming schema and have a "replica A"
and a "replica B".

This still hardcodes the "codfw" part and the host names but I am
trying to not change multiple things at once and keep it simple
in each patch.

So this should simply start replicating to the new host and it
slightly renames the existing replication config snippet.

Bug: T387833
Bug: T372804
Hosts: O:gerrit
Change-Id: Iff23a4d3ff190cbe94f6dfc178a22e998136e0f4