wmflib: get_lvs_class_hosts() to exclude test LVS hosts
We specify lvs1013 as secondary LVS, which it is, but it is for testing
purposes and so should not be returned when the cumin alias
A:lvs-secondary-eqiad is queried for.

    sukhe@cumin1002:~$ sudo cumin "A:lvs-secondary-eqiad"
    2 hosts will be targeted:
    lvs[1013,1020].eqiad.wmnet
    DRY-RUN mode enabled, aborting

This is somewhat manual for now where we exclude hosts specifically but
that's fine -- an automated solution is unlikely anyway since we can't
exclude by role as well (LVS vs Liberica).

Hosts: cumin1002.eqiad.wmnet
Change-Id: I6f667ce8251ee2653d46b9ef769fc7513730cb85