Update dumpsgen SSH settings on clouddumps servers.
We are beginning to sync the dumps from the cephfs volume to the
clouddumps servers. At the moment the destionation directory is separate
from /srv/dumps so that we don't overwrite the current production dumps.

We configure the SSH daemon to chroot connections from the dumpsgen
user to the /srv/mediawiki-dumps-legacy directory.

For now, we are removing the ForceCommand setting from the
authorized_keys file, while we begin using rsync from the sync pods.

Hosts: clouddumps1001.wikimedia.org,clouddumps1002.wikimedia.org
Bug: T389784
Change-Id: Id22b031d10f3b3826110b14819762c38dc8b83f3