Extend package list to be installed from component/puppet7 on trixie
To unbreak some dependency cycles, ruby-defaults (which is the package which
declares what version of rubyX.Y is pulled in by ruby etc. meta packages)
declares a Breaks: on the puppet7-agent. This blocks deploying the Puppet
Puppet 7 agent and we don't need it for our upgrades, so ruby-defaults
was rebuilt without the Breaks: statement

Bug: T392790
Change-Id: I82366f2986818dcdfc67863c55f4b2940302a442