systemd::sysuser: create the user synchronously in the define
Prior to this change the user was created with the systemd-sysusers
command using an exec with the refreshonly flag. The exec was triggered
from a notify on the sysusers config. Notifies are not triggered
immediately by Puppet, which means the user is not immediately available
for subsequent puppet resources, such as creating a file with the
defined user, resulting in Puppet failures.

After this change we run the exec to create the user on every puppet
run, only if systemd-sysusers reports in its dry-run mode that changes
would be made.

The dry-run is only supported on bookworm and newer, so use the old
logic for bullseye and earlier releases.

Hosts: C:certspotter
Hosts: C:external_clouds_vendors
Hosts: C:gitlab
Hosts: C:mediabackup::storage
Hosts: C:mediabackup::worker
Hosts: C:phabricator::phd
Hosts: P:vopsbot
Change-Id: Ib2b0509e16a49b211377a775fd60ee6c1684b4e4