raid: update facter and get-raid-status to allow storcli
Update the get-raid-status-perccli script to allow the usage
of storcli, a Broadcom tool that shares a lot of commands with
perccli. This is needed since the Supermicro Config J/E servers
run the SAS 3908 controller, but perccli seems not able to detect
all the info that we need.
Add also support for the following use cases:
- No BBU on the controller (some ms-be nodes using JBOD).
- Controller 38XX (Passthrough controller for JBOD).

Bug: T393146
Change-Id: I32da4235d88612fb3b76161c3bd70c9eaa1c191c