benthos/mw_accesslog_metrics: increase buffering
Benthos started lagging for unclear reasons, so increase buffering to
help it cope.

Because it should be moved to k8s and I have a half-baked CR on my
radar, I do not think an in depth investigation of why this is
happening all of a sudden is worth the trouble.

Change-Id: I3deb989cfc92ab5e863a5660b1443821077af7d8