JBOD partman recipe for Cassandra
There are a number of custom Cassandra partman recipes, this is meant
to eventually succeed/replace them (which assumes a move toward JBOD
as a standard).

It builds on standard-efi.cfg and creates a raid of approx. 60G, which
hosts volumes for swap, /, and /srv/cassandra/instance (common storage
for commitlogs, hints, cache, etc).  The remaining space is allocated
for ext4 filesystems; These are the 'D's in JBOD (aka they will,
comprise the `data_file_directories` set).

This is meant to be used in conjunction with raid1-2dev-efi.cfg,
creating a RAID1 across the first two devices.  The assumption is that
handing of the data partitions, including those on any devices beyond
the first two (`/dev/sd[c-z]`), would be left as a post-install
exercise (ideally a cookbook).

Bug: T391544
Change-Id: If630a2f274217f4ccae3590df4ac5a532d182957