run_ci_locally.sh: use bind mounts for local runs
Prior to this change, for local CI runs, the local src repo was
tarred and extracted onto the container src repo. This has
two disadvantages:

1. The tar is slow, ~15secs
2. New edits to the source repo are not propagated in the container, so
   you can't re-run tests, without restarting the container.

After this change we mount the src repo's files as readonly on top of
the container's src, except for the test dependencies, which solves both
negatives.

Change-Id: I3c765f0c1ae3d987a7cec598bed7f1e684bc506b