P:mw::maintenance::refreshlinks: rename and prepare for mw-cron
Currently, profile::mediawiki::maintenance::refreshlinks does not use
sharded_periodic_job, instead relying on a specialized job resource
definition that still supports the '<shard>@<day-of-month>' idiom.

While it would be "easy" to revive support for this deprecated syntax in
sharded_periodic_job and then adopt it here, that's non-ideal, as we do
not want to encourage its use more broadly.

At the same time, it would also be non-ideal to retain the specialized
definition, as it would require plumbing various mw-cron params down to
the profile::mediawiki::periodic_job instantiation therein and still use
the '<shard>@<day-of-month>' idiom (meh).

Instead, this deprecates the specialized job definition, in favor of
defining shard-to-day-of-month mapping in a more structured way. This
complexity is limited to profile::mediawiki::maintenance::refreshlinks,
rather than pushing it into sharded_periodic_job.

While doing so, we also:
1. Imbue the jobs with a far more descriptive name going forward, as
   already discussed in T388530.
2. Prepare for migration to mw-cron (e.g., adding relevant periodic_job
   resource params).

While this patch does change the set of puppet resources, it does not
make system-affecting changes (nor are any on-host resources abandoned).

Hosts: mwmaint1002.eqiad.wmnet, deploy1003.eqiad.wmnet
Bug: T388530
Change-Id: I453b09e7767082a2466c129ca63dcf1598dff13a