P:mw::maintenance::refreshlinks: migrate s6 to mw-cron
This pilots migration on s6, which is a (comparatively) low-impact and
fast (i.e., completes quickly) shard to test on.

The assumption here is that a manual re-run of the s6 job (e.g., to
verify consistent output with the recent run on May 6th) is safe.

Hosts: mwmaint1002.eqiad.wmnet, deploy1003.eqiad.wmnet
Bug: T388530
Change-Id: Id6415fc3f475d0df9439a4b045d93244d17ed3db