pdb_resource_exporter: add puppetdb resource exporter to puppedb
pdb_resource_exporter is a script that queries PuppetDB, retrieves
the cardinality of a resource, and pushes the result to the
Prometheus Pushgateway.
A jq filter can be applied to each query to refine the results.
Multiple queries can be defined.
Each query will produce a series of the puppetdb_resource_stats
metric.

Bug: T395442
Change-Id: I1f4f1957ba12fa9d26405b741e52aff92e50cb6f