varnish: Allow customising "contact noc@" error
Extracts the "contact noc@wikimedia.org" part of error messages to a
vcl_config entry, with the current value used as a default. This is to
allow deployment-prep to have it point somewhere else instead.

The verb is included in the parameter to support usage proposed in [0]
with a wiki page ('see https://...') instead of a plain email address.

[0]: https://phabricator.wikimedia.org/T393404#10802746

Bug: T393404
Hosts: cp2027.codfw.wmnet,cp6001.drmrs.wmnet
Change-Id: Ieb29b112b8446a166ba9c367cd6558cfd830c0d7