trafficserver: route a smaller subset of enwiki pcs pages without restbase
Given the overall size of traffic for enwiki against mobileapps/pcs
endpoints, there are concerns about cache hit rate and traffic impact
on the rest-gateway and on the mobileapps service itself. Make the
ignore rules for enwiki slightly more specific, effectively allowing
all pages starting with A to be routed without using restbase. This is
in anticipation of making larger swathes of characters routed directly
in later changes.

Bug: T393591
Change-Id: I938d82cf65531853ba3a627ce9ddf26a0b11c718