Update the managed airflow temp directory for dumps on k8s
We are currently creating dumps in parallel from both the snapshot
servers and from an Airflow instance.

The files being published by the new Airflow based pipelines were
writing to /srv/mediawiki-dumps-legacy as a temporary directory, but it
turns out that this is on the root volume.

This commit creates a new directory for these temporary dumps files,
alongside the production dumps on the big /srv/dumps volume.

We will need a corresponding change in the Airflow dumps DAG to use this
directory.

Hosts: O:dumps::distribution::server
Bug: T389784
Change-Id: I5642ed7c4803709b0196074ea621d9d2b39b26e9