cirrussearch: add cirrussearch row A/remove elastic row B
As part of the OpenSearch migration, we need to remove the
old hosts from load balancer pools so we don't cause LVS
alerts. We also need to add the new hosts so they will be
able to receive search traffic.

Additionally, remove the row A masters with the outdated
'elastic' names and add the row B masters with the new
'cirrussearch' names. A subsequent patch will remove the
outdated row B master names after we finish with the
row B reimages.

Hosts: elastic1085.eqiad.wmnet,cirrussearch1112.eqiad.wmnet
Bug: T391118
Bug: T388610
Change-Id: Ia20000b9409b0025fcf3ad6a368bb8b64b34aed3