clouddumps: Manage directories beneath /srv/dumps/xmldatadumps_airflow_temp
At the moment, we are syncing to Airflow dump jobs and index files to a
temporary location on the clouddumps servers, which is a subdirectory of
/srv/dumps

The jobs have been failing recently because the xmldatadumps/public
subdirectory doesn't exist.

This commit adds those two additional temporary managed directories.

Hosts: O:dumps::distribution::server
Bug: T389784
Change-Id: I404127dc625643730e12785ca9e89e6a7e0fe8ec