mw::maintenance: Run listTaskCounts every 2 hours
Why:
The former systemd timers didn't replace the currently running script if
it was still running. However, in mw-cron on kubernetes, the default
concurrencyPolicy is set to Replace, which terminates the running
script, very probably causing the errors reported in T394018 and
T394019.

What:
Set listTaskCounts to run every 2 hours so it has time to finish.

Hosts: deploy1003.eqiad.wmnet
Bug: T394018
Bug: T394019
Change-Id: Ie7615e80d4bca3221556b76e80cee66813bb40ca