aptrepo: Provide thirdparty/docker component with upstream packages
Establish thirdparty/docker component with upstream Docker packages for
production hosts that require newer versions of Docker tooling than
those provided by official Debian packages.

Having a component independent of thirdparty/ci makes it clear that
these packages are not solely intended for use in CI but all hosts that
require newer Docker packages such as deploy hosts, and follows the
convention of other components named after specific software.

A component for trixie is included because unfortunately the version of
docker-buildx in trixie is already quite old and `docker buildx` is the
recommended CLI for building Docker images. (The "legacy" `docker build`
builder has been deprecated for some time.)

The need for `docker buildx` on deploy hosts is the main driver of this
change (see T392526). However, the availability of newer Docker packages
in bookworm (and perhaps trixie) should benefit other projects such as
the upcoming Zuul migration.

Bug: T392526
Change-Id: Id29aed8347dc49e4aa5a4ffecbc4c419ce00767d