hdfs: add an-worker1177 to in retup role
an-worker1177 recently had its harddrives swapped but is currently in a read-only state which
does not allow for much even on trying to reboot it.
This patch adds the host to the insetup role to allow us to run puppet when the host is reimaged
and also adds the host to use the analytics-flex.cfg recipe used for the an-worker hosts.

Bug: T390171
Change-Id: I30610727ab30617f5e21bdfd2cba215d0630223f