Remove krb1001 from the list of KDCs presented to clients
If one attempts to change a password, it happens on krb1002 (the admin
server), but with krb1001 still being present in the krb5.conf, the
Kerberos clients will initially attempt to contact krb1001. But given
kpropd only runs every hour, these password changes are otherwise
ineffective.

As part of the procedure to phase out krb1001, drop it from the KDC
list used on clients.

Bug: T390863
Change-Id: I6817e62238986a017b459e7c83639bb1d4c9486c