P:wmcs: cloudgw: Remove internet access for octavia-lb-mgmt-net
We do not want the Octavia management network to be able to egress
traffic to the internet. To do that, remove it from the list of networks
that egress NAT is applied to; this will cause router filters to reject
the traffic (as it's coming from a private address) like the v6 net
already will.

Bug: T394099
Change-Id: I4ce8b5dfef63b586f7c83a2ace790c6351c632ee